everify/requirements.txt
requests>=2.25.0
PyJWT>=2.0.0
psycopg2-binary>=2.9.0 